+++ src/core/conversions/swap.ml
@@ -30,6 +30,7 @@ class swap ~field (source : source) =
     method remaining = source#remaining
     method abort_track = source#abort_track
     method seek = source#seek
+    method seek_source = source
     method self_sync = source#self_sync
 
     method private get_frame buf =