+++ src/core/operators/available.ml
@@ -29,6 +29,7 @@ class available ~track_sensitive ~override p (source : source) =
     method remaining = source#remaining
     method abort_track = source#abort_track
     method seek = source#seek
+    method seek_source = source
     method self_sync = source#self_sync
     val mutable ready = p ()