+++ src/core/operators/chord.ml
@@ -51,6 +51,7 @@ class chord metadata_name (source : source) =
     method is_ready = source#is_ready
     method abort_track = source#abort_track
     method seek = source#seek
+    method seek_source = source
     method self_sync = source#self_sync
     val mutable notes_on = []