+++ src/core/operators/compress_exp.ml
@@ -30,6 +30,7 @@ class compress ~field (source : source) mu =
     method is_ready = source#is_ready
     method abort_track = source#abort_track
     method seek = source#seek
+    method seek_source = source
     method self_sync = source#self_sync
 
     method private get_frame buf =