+++ src/core/operators/delay_line.ml
@@ -33,6 +33,7 @@ class delay (source : source) duration =
     method remaining = source#remaining
     method abort_track = source#abort_track
     method seek = source#seek
+    method seek_source = source
     method self_sync = source#self_sync
 
     (** Length of the buffer in samples. *)