+++ src/core/operators/fir_filter.ml
@@ -125,6 +125,7 @@ class fir (source : source) freq beta numcoeffs =
     method remaining = source#remaining
     method is_ready = source#is_ready
     method seek = source#seek
+    method seek_source = source
     method abort_track = source#abort_track
     method self_sync = source#self_sync