+++ src/core/operators/lufs.ml
@@ -110,6 +110,7 @@ class lufs window source =
     method is_ready = source#is_ready
     method remaining = source#remaining
     method seek = source#seek
+    method seek_source = source
     method abort_track = source#abort_track
     method self_sync = source#self_sync
     method channels = self#audio_channels