+++ src/core/operators/on_metadata.ml
@@ -28,6 +28,7 @@ class on_metadata f s =
     method abort_track = s#abort_track
     method remaining = s#remaining
     method seek n = s#seek n
+    method seek_source = s
     method self_sync = s#self_sync
 
     method private get_frame ab =