+++ src/core/operators/still_frame.ml
@@ -32,6 +32,7 @@ class still_frame ~name (source : source) =
     method stype = source#stype
     method remaining = source#remaining
     method seek = source#seek
+    method seek_source = source
     method self_sync = source#self_sync
     method is_ready = source#is_ready
     method abort_track = source#abort_track