+++ src/core/synth/keyboard_sdl.ml
@@ -87,6 +87,7 @@ class keyboard velocity =
   object (self)
     inherit Source.active_source ~name:"input.keyboard.sdl" ()
     inherit Source.no_seek
+    method seek_source = (self :> Source.source)
     method stype = `Infallible
     method is_ready = true
     method remaining = -1